Definitive version without router.
