+++ package.json
@@ -19,7 +19,6 @@
     "pako": "^1.0.6",
     "truffle-contract": "^3.0.0",
     "vue": "^2.5.16",
-    "vue-router": "^2.7.0",
     "vuetify": "^1.1.7"
   },
   "devDependencies": {
