--- src/router/index.js
@@ -1,15 +1,0 @@
-import Vue from 'vue'
-import Router from 'vue-router'
-import Dashboard from '@/components/Dashboard'
-
-Vue.use(Router)
-
-export default new Router({
-  routes: [
-    {
-      path: '/',
-      name: 'dashboard',
-      component: Dashboard
-    }
-  ]
-})
